跑通了一个注册模块，实现了再js前端对密码进行md5加密的功能，计划在后台再实现一次，但是感觉wangEditor富文本编辑器没有其他的好使，之后在尝试一次试试

+++ myBlog/src/main/java/com/yuan/model/blogComment.java
@@ -5,7 +5,7 @@ import java.sql.Timestamp;
 
 public class blogComment {
     private int commentId;
-    private int blogId;
+    private long blogId;
     private String blogComment;
     private int commentStatus;
     private Timestamp buildTime;
@@ -18,11 +18,11 @@ public class blogComment {
         this.commentId = commentId;
     }
 
-    public int getBlogId() {
+    public long getBlogId() {
         return blogId;
     }
 
-    public void setBlogId(int blogId) {
+    public void setBlogId(long blogId) {
         this.blogId = blogId;
     }
